added validations to prompt and genereated a README.md
