+++ Develop/README.md
@@ -1,0 +1,54 @@
+
+  
+[![MIT license](https://img.shields.io/badge/License-MIT-blue.svg)](https://lbesson.mit-license.org/)
+# README GENERATOR
+## Table of Contents
+- README GENERATOR
+- Description
+- Installation
+- Usage
+- License
+- Contributing
+- Tests
+- Questions
+  - Stevomunoz@yahoo.com
+## Description
+This application is meant for users to generate a README.md file by using a template that uses command line interface to ask user questions regarding their application created.
+## Installation
+step 1: clone https://github.com/Steve-Munoz/README-GENERATOR.git 
+step 2: install node https://nodejs.org/en/download/ 
+step 3: npm install 
+step 4: node readMe.js
+## Usage
+Easily Generate a README.md file
+## License
+MIT License
+
+Copyright (c) [2020] [Steve Munoz]
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+## Contributing
+Steve Munoz
+## Tests
+no tests were performed for this application
+## Questions
+### User GitHub profile picture
+- <img src = "images/GitHub-pic.jpg" width = "150">
+### User GitHub email
+- Stevomunoz@yahoo.com
